chore: refactored handle-proof-syntax-check.ts into subfunctions into dir/file hierarchy
